# [sdk] SuppressInstrumentationScope perf tweaks