--- a/src/OpenTelemetry/SuppressInstrumentationScope.cs
+++ b/src/OpenTelemetry/SuppressInstrumentationScope.cs
@@ -25,8 +25,8 @@ namespace OpenTelemetry
     {
         // An integer value which controls whether instrumentation should be suppressed (disabled).
         // * null: instrumentation is not suppressed
-        // * Value = [int.MinValue, -1]: instrumentation is always suppressed
-        // * Value = [1, int.MaxValue]: instrumentation is suppressed in a reference-counting mode
+        // * Depth = [int.MinValue, -1]: instrumentation is always suppressed
+        // * Depth = [1, int.MaxValue]: instrumentation is suppressed in a reference-counting mode
         private static readonly RuntimeContextSlot<SuppressInstrumentationScope?> Slot = RuntimeContext.RegisterSlot<SuppressInstrumentationScope?>("otel.suppress_instrumentation");
 
         private readonly SuppressInstrumentationScope? previousScope;