--- a/test/Benchmarks/SuppressInstrumentationScopeBenchmarks.cs
+++ b/test/Benchmarks/SuppressInstrumentationScopeBenchmarks.cs
@@ -1,0 +1,44 @@
+// <copyright file="SuppressInstrumentationScopeBenchmarks.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using BenchmarkDotNet.Attributes;
+
+namespace OpenTelemetry.Benchmarks;
+
+public class SuppressInstrumentationScopeBenchmarks
+{
+    [Benchmark]
+    public void Begin()
+    {
+        using var scope1 = SuppressInstrumentationScope.Begin();
+
+        using var scope2 = SuppressInstrumentationScope.Begin();
+
+        using var scope3 = SuppressInstrumentationScope.Begin();
+    }
+
+    [Benchmark]
+    public void Enter()
+    {
+        SuppressInstrumentationScope.Enter();
+
+        SuppressInstrumentationScope.IncrementIfTriggered();
+
+        SuppressInstrumentationScope.DecrementIfTriggered();
+
+        SuppressInstrumentationScope.DecrementIfTriggered();
+    }
+}