## Changes

Tweaked the implementation of `SuppressInstrumentationScope` to achieve better perf.

Before...

* We stored an `int` in the context slot. That `int` had to be boxed when scopes were created and we boxed a 0 when scopes were disposed.
* Any time we incremented or decremented in reference counting mode we pushed a new `int` to the context slot which had to be boxed AND creates a new `ExecutionContext`.

After...

* We store the `SuppressInstrumentationScope` instance itself in the context slot. That avoids the boxing and gives us storage for the reference count which we can increment/decrement freely without creating any new `ExecutionContext`s.

## Benchmarks

Before:

| Method |      Mean |    Error |   StdDev |   Gen0 | Allocated |
|------- |----------:|---------:|---------:|-------:|----------:|
|  Begin | 143.91 ns | 2.027 ns | 1.896 ns | 0.0515 |     648 B |
|  Enter |  99.49 ns | 1.226 ns | 1.147 ns | 0.0305 |     384 B |

After:

| Method |      Mean |    Error |   StdDev |   Gen0 | Allocated |
|------- |----------:|---------:|---------:|-------:|----------:|
|  Begin | 125.37 ns | 1.738 ns | 1.626 ns | 0.0362 |     456 B |
|  Enter |  59.17 ns | 0.710 ns | 0.664 ns | 0.0082 |     104 B |

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (nullable enabled, static analysis, etc.)
* [X] Appropriate `CHANGELOG.md` files updated for non-trivial changes
* [X] Unit tests added/updated